Broke DrawerController and MainPagerAdapterComment classes out of MainActivity.java
